backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
rq==1.16.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
pgvector==0.2.5
aiofiles==23.2.1
